+++ packages/flutter_tools/test/commands/run_test.dart
@@ -3,19 +3,13 @@
 // found in the LICENSE file.
 
 import 'package:flutter_tools/src/base/common.dart';
-import 'package:flutter_tools/src/build_info.dart';
 import 'package:flutter_tools/src/commands/run.dart';
-import 'package:flutter_tools/src/device.dart';
-import 'package:flutter_tools/src/runner/flutter_command.dart';
-import 'package:mockito/mockito.dart';
 
 import '../src/common.dart';
 import '../src/context.dart';
 import '../src/mocks.dart';
 
 void main() {
-  final MockDeviceManager mockDeviceManager = MockDeviceManager();
-
   group('run', () {
     testUsingContext('fails when target not found', () async {
       final RunCommand command = RunCommand();
@@ -27,65 +21,5 @@ void main() {
         expect(e.exitCode ?? 1, 1);
       }
     });
-
-    testUsingContext('should only request artifacts corresponding to connected devices', () async {
-      when(mockDeviceManager.getDevices()).thenAnswer((Invocation invocation) {
-        return Stream<Device>.fromIterable(<Device>[
-          MockDevice(TargetPlatform.android_arm),
-        ]);
-      });
-
-      expect(await RunCommand().requiredArtifacts, unorderedEquals(<DevelopmentArtifact>{
-        DevelopmentArtifact.universal,
-        DevelopmentArtifact.android,
-      }));
-
-      when(mockDeviceManager.getDevices()).thenAnswer((Invocation invocation) {
-        return Stream<Device>.fromIterable(<Device>[
-          MockDevice(TargetPlatform.ios),
-        ]);
-      });
-
-      expect(await RunCommand().requiredArtifacts, unorderedEquals(<DevelopmentArtifact>{
-        DevelopmentArtifact.universal,
-        DevelopmentArtifact.iOS,
-      }));
-
-      when(mockDeviceManager.getDevices()).thenAnswer((Invocation invocation) {
-        return Stream<Device>.fromIterable(<Device>[
-          MockDevice(TargetPlatform.ios),
-          MockDevice(TargetPlatform.android_arm),
-        ]);
-      });
-
-      expect(await RunCommand().requiredArtifacts, unorderedEquals(<DevelopmentArtifact>{
-        DevelopmentArtifact.universal,
-        DevelopmentArtifact.iOS,
-        DevelopmentArtifact.android,
-      }));
-
-      when(mockDeviceManager.getDevices()).thenAnswer((Invocation invocation) {
-        return Stream<Device>.fromIterable(<Device>[
-          MockDevice(TargetPlatform.web),
-        ]);
-      });
-
-      expect(await RunCommand().requiredArtifacts, unorderedEquals(<DevelopmentArtifact>{
-        DevelopmentArtifact.universal,
-        DevelopmentArtifact.web,
-      }));
-    }, overrides: <Type, Generator>{
-      DeviceManager: () => mockDeviceManager,
-    });
   });
-}
-
-class MockDeviceManager extends Mock implements DeviceManager {}
-class MockDevice extends Mock implements Device {
-  MockDevice(this._targetPlatform);
-
-  final TargetPlatform _targetPlatform;
-
-  @override
-  Future<TargetPlatform> get targetPlatform async => _targetPlatform;
 }
